Update email features.
