+++ hamail.php
@@ -5,7 +5,6 @@ Plugin URI: https://wordpress.org/plugins/hamail/
 Description: A WordPress plugin for sending e-mail via SendGrid.
 Author: Hametuha INC.
 Version: nightly
-PHP Version: 5.6
 Author URI: https://hametuha.co.jp/
 License: GPL3 or later
 License URI: https://www.gnu.org/licenses/gpl-3.0.html
@@ -50,7 +49,7 @@ function hamail_plugins_loaded( $plugin ) {
 		$auto_loader = __DIR__ . '/vendor/autoload.php';
 		if ( ! file_exists( $auto_loader ) ) {
 			// translators: %s is composer path.
-			throw new Exception( sprintf( __( '[Hamail] PHP auto loader %s is missing. Did you run <code>composer install</code>?', 'hamail' ), $auto_loader ) );
+			throw new Exception( sprintf( __( '[Hamail] PHP autoloader %s is missing. Did you run <code>composer install</code>?', 'hamail' ), $auto_loader ) );
 		}
 		require $auto_loader;
 		// Load functions.
@@ -62,10 +61,14 @@ function hamail_plugins_loaded( $plugin ) {
 				}
 			}
 		}
+		// Transaction mail Sender
+		\Hametuha\Hamail\API\TransactionMails::get_instance();
+		// Setting Screen
+		Hametuha\Hamail\Ui\SettingsScreen::get_instance();
 		// Dynamic emails.
 		Hametuha\Hamail\API\DynamicEmails::get_instance();
 		// SMTP Handlers
-		\Hametuha\Hamail\Controller\SmtpController::get_instance();
+		Hametuha\Hamail\Controller\SmtpController::get_instance();
 		// Screens.
 		Hametuha\Hamail\Ui\ListTable\RecipientsColumn::get_instance();
 		// Enable user sync.
